# api/openapi.yaml
openapi: 3.0.3
info:
  title: Workouts Explorer API
  version: 0.2.0
  description: Design-first API for the secure account lifecycle.
servers:
  - url: /
paths:
  /api/config:
    get:
      operationId: getPublicConfig
      tags: [public]
      responses:
        '200': {description: Safe browser configuration, content: {application/json: {schema: {$ref: '#/components/schemas/PublicConfig'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /api/openapi.yaml:
    get:
      operationId: getOpenAPIDocument
      tags: [public]
      responses:
        '200': {description: OpenAPI YAML, content: {application/yaml: {schema: {type: object, additionalProperties: true}}}}
        default: {$ref: '#/components/responses/Problem'}
  /swagger:
    get:
      operationId: getSwaggerUI
      tags: [public]
      responses:
        '200': {description: Swagger UI, content: {text/html: {schema: {type: string}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
    cookieAuth: {type: apiKey, in: cookie, name: workouts_session}
...
  responses:
    Problem:
      description: RFC 9457 Problem Details
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
...
  schemas:
    PublicConfig:
      type: object
      additionalProperties: false
      required: [productName, pollingIntervalSeconds, mapFitPaddingPixels, passwordMinimumLength, pageSizeMaximum]
      properties:
        productName: {type: string}
        pollingIntervalSeconds: {type: integer, minimum: 1, maximum: 3600}
        mapFitPaddingPixels: {type: integer, minimum: 0, maximum: 512}
        passwordMinimumLength: {type: integer, minimum: 12, maximum: 64}
        pageSizeMaximum: {type: integer, minimum: 25, maximum: 1000}
        baseMapTileUrl: {type: string, format: uri-template}
        baseMapAttribution: {type: string, maxLength: 1024}
...
    ValidationError:
      type: object
      additionalProperties: false
      required: [field, code]
      properties:
        field: {type: string}
        code: {type: string}
        message: {type: string}
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, requestId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        requestId: {type: string}
        errors: {type: array, items: {$ref: '#/components/schemas/ValidationError'}}
...